+++ HISTORY.rst
@@ -5,6 +5,11 @@ History
 X.X.X (XXXX-XX-XX)
 ------------------
 
+* Update the required version of pika to be >=1.0.0 and update the RMQComm/RMQAsyncComm code to use the updated API
+
+X.X.X (XXXX-XX-XX)
+------------------
+
 * Added C, C++, Fortran, Matlab, R versions of server in rpc_lesson1 example
 * Added C, C++, Fortran, Matlab, R versions of server in rpc_lesson2 example
 * Added C, C++, Fortran, Matlab, R versions of server in rpc_lesson2b example